\name{nopp-package}
\alias{nopp-package}
\alias{nopp}
\docType{package}
\title{
Nash Optimal Party Positions
}
\description{
Estimation of party/candidate ideological positions that correspond to a Nash equilibrium along a one-dimensional space}
\details{
\tabular{ll}{
Package: \tab nopp\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2012-06-26\cr
License: \tab GPL (>= 2)\cr
}
\code{nopp} is a package for \code{R} which enables to compute party/candidate ideological positions that correspond to a Nash Equilibrium along a one-dimensional space. It accommodates alternative motivations in (each) party strategy while allowing to estimate the uncertainty around their optimal positions through two different procedures (bootstrap and MC).
}
\author{
Luigi Curini, Stefano M. Iacus

Maintainer: Luigi Curini <luigi.curini@unimi.it>, Stefano M. Iacus <stefano.iacus@unimi.it>
}
\references{
 Adams, James F., Samuel Merrill III, and Bernard Grofman (2005). A Unified Theory of Party Competition. Cambridge: Cambridge University Press
 
 Merrill, Samuel III, and James Adams (2001), Computing Nash Equilibria in Probabilistic, Multiparty Spatial Models with Nonpolicy Components, Political Analysis, 9, 347--61

Curini, L., and Iacus, S.M. (2017) Nash Optimal Party Positions: The nopp R Package, Journal of Statistical Software, 81(11), 1--25
}
\keyword{ package }
